Added comments
Added comments and changed currencyconversion rate typing to __int64

+++ dZoneManager/dZoneManager.h
@@ -34,7 +34,7 @@ public:
 	void AddSpawner(LWOOBJID id, Spawner* spawner);
 	LWOZONEID GetZoneID() const;
 	uint32_t GetMaxLevel();
-	uint32_t GetLevelCapCurrencyConversion();
+	__int64 GetLevelCapCurrencyConversion();
 	LWOOBJID MakeSpawner(SpawnerInfo info);
 	Spawner* GetSpawner(LWOOBJID id);
 	void RemoveSpawner(LWOOBJID id);
@@ -44,8 +44,16 @@ public:
 	Entity* GetZoneControlObject() { return m_ZoneControlObject; }
 
 private:
+	/**
+	 * The maximum level of the world.
+	 */
 	uint32_t m_MaxLevel = 0;
-	uint32_t m_CurrencyConversionRate = 0;
+
+	/**
+	 * The ratio of LEGO Score to currency when the character has hit the max level.
+	 */
+	__int64 m_CurrencyConversionRate = 0;
+	
     static dZoneManager* m_Address; //Singleton
 	Zone* m_pZone;
 	LWOZONEID m_ZoneID;
